+++ agrest-engine/src/main/java/io/agrest/runtime/DefaultUpdateBuilder.java
@@ -129,7 +129,8 @@ public class DefaultUpdateBuilder<T> implements UpdateBuilder<T> {
      */
     @Override
     public SimpleResponse sync(EntityUpdate<T> update) {
-        return sync(Collections.singleton(update));
+        Collection<EntityUpdate<T>> updates = update != null ? Collections.singleton(update) : Collections.emptyList();
+        return sync(updates);
     }
 
     /**
@@ -164,7 +165,8 @@ public class DefaultUpdateBuilder<T> implements UpdateBuilder<T> {
      */
     @Override
     public DataResponse<T> syncAndSelect(EntityUpdate<T> update) {
-        return syncAndSelect(Collections.singleton(update));
+        Collection<EntityUpdate<T>> updates = update != null ? Collections.singleton(update) : Collections.emptyList();
+        return syncAndSelect(updates);
     }
 
     private SimpleResponse doSync() {