+++ agrest-jaxrs3/src/main/java/io/agrest/jaxrs3/provider/EntityUpdateReader.java
@@ -14,7 +14,6 @@ import jakarta.ws.rs.core.MultivaluedMap;
 import jakarta.ws.rs.ext.MessageBodyReader;
 import jakarta.ws.rs.ext.Provider;
 
-import java.io.IOException;
 import java.io.InputStream;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
@@ -22,14 +21,12 @@ import java.util.Collection;
 
 /**
  * A provider of {@link MessageBodyReader} for {@link EntityUpdate} parameters.
- *
- * @since 1.20
  */
 @Provider
 @Consumes(MediaType.APPLICATION_JSON)
 public class EntityUpdateReader implements MessageBodyReader<EntityUpdate<?>> {
 
-    private EntityUpdateReaderProcessor reader;
+    private final EntityUpdateReaderProcessor reader;
 
     public EntityUpdateReader(@Context Configuration config) {
         this.reader = new EntityUpdateReaderProcessor(
@@ -43,14 +40,19 @@ public class EntityUpdateReader implements MessageBodyReader<EntityUpdate<?>> {
     }
 
     @Override
-    public EntityUpdate<?> readFrom(Class<EntityUpdate<?>> type, Type genericType, Annotation[] annotations,
-                                    MediaType mediaType, MultivaluedMap<String, String> httpHeaders, InputStream entityStream)
-            throws IOException, WebApplicationException {
+    public EntityUpdate<?> readFrom(
+            Class<EntityUpdate<?>> type,
+            Type genericType,
+            Annotation[] annotations,
+            MediaType mediaType,
+            MultivaluedMap<String, String> httpHeaders,
+            InputStream entityStream) throws WebApplicationException {
 
         Collection<EntityUpdate<Object>> updates = reader.read(genericType, entityStream);
 
         if (updates.isEmpty()) {
-            throw AgException.badRequest("No update");
+            // null is valid here
+            return null;
         }
 
         if (updates.size() > 1) {